+++ wcomponents-core/src/main/java/com/github/bordertech/wcomponents/WMultiFileWidget.java
@@ -739,16 +739,13 @@ public class WMultiFileWidget extends AbstractInput implements Targetable, AjaxI
 		// Wrap the file item
 		FileItemWrap wrap = new FileItemWrap(items[0]);
 
-		// if fileType is supplied then validate it
-		if (hasFileTypes() && !FileUtil.validateFileType(wrap, getFileTypes())) {
-			String invalidMessage = FileUtil.getInvalidFileTypeMessage(getFileTypes());
-			throw new SystemException(invalidMessage);
+		// Validate the file type
+		if (hasFileTypes()) {
+			doHandleUploadedFileTypeValidation(wrap);
 		}
-
-		// if fileSize is supplied then validate it
-		if (hasMaxFileSize() && !FileUtil.validateFileSize(wrap, getMaxFileSize())) {
-			String invalidMessage = FileUtil.getInvalidFileSizeMessage(getMaxFileSize());
-			throw new SystemException(invalidMessage);
+		// Validate the file size
+		if (hasMaxFileSize()) {
+			doHandleUploadedFileSizeValidation(wrap);
 		}
 
 		FileWidgetUpload file = new FileWidgetUpload(fileId, wrap);
@@ -757,6 +754,30 @@ public class WMultiFileWidget extends AbstractInput implements Targetable, AjaxI
 		// Set the file id to be used ion the renderer
 		setFileUploadRequestId(fileId);
 		setNewUpload(true);
+	}
+
+	/**
+	 * Perform file type validation on the uploaded file.
+	 *
+	 * @param wrap the file item to validate
+	 */
+	protected void doHandleUploadedFileTypeValidation(final FileItemWrap wrap) {
+		if (!FileUtil.validateFileType(wrap, getFileTypes())) {
+			String invalidMessage = FileUtil.getInvalidFileTypeMessage(getFileTypes());
+			throw new SystemException(invalidMessage);
+		}
+	}
+
+	/**
+	 * Perform file size validation on the uploaded file.
+	 *
+	 * @param wrap the file item to validate
+	 */
+	protected void doHandleUploadedFileSizeValidation(final FileItemWrap wrap) {
+		if (!FileUtil.validateFileSize(wrap, getMaxFileSize())) {
+			String invalidMessage = FileUtil.getInvalidFileSizeMessage(getMaxFileSize());
+			throw new SystemException(invalidMessage);
+		}
 	}
 
 	/**